requirements.txt
# Core dependencies for multimodal video transcription system

# Audio/Video Processing
sarvamai
requests>=2.31.0
ffmpeg-python==0.2.0
opencv-python==4.8.1.78
librosa==0.10.1
soundfile==0.12.1

# Speaker Diarization (clustering-based, no external dependencies)
scikit-learn>=1.3.0

# Deep Learning
torch==2.1.0
torchaudio==2.1.0
numpy==1.24.3

# Facial Emotion Detection
deepface==0.0.79
tensorflow

# LLM API
groq

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
